Make http mcp a separate project so that local mcp only depends dotnet runtime (no more aspnetcore dependencies)
